refactoring code part 2
