+++ examples/cli-langgraph-agents/demo.py
@@ -1,43 +1,45 @@
 import os
 from dotenv import load_dotenv
-
+load_dotenv()
+from langgraph.prebuilt import create_react_agent
+from langgraph_supervisor import create_supervisor
+from langchain_openai import ChatOpenAI
 from ak import CLI
 from ak_langgraph import AgentModule, AgentSession
-
-from langgraph_custom_agent import LangGraphAgent, LiteLLMModel
-
-load_dotenv()
+from custom_langgraph_agent import CustomLangGraphAgent
 
 session = AgentSession()
 
-llm = LiteLLMModel(model_name="gemini/gemini-2.5-flash", temperature=0.0, api_key=os.getenv("GEMINI_API_KEY"))
+model = ChatOpenAI(model_name="gpt-4o-mini", temperature=0.0, openai_api_key=os.getenv("OPENAI_API_KEY"))
 
-math_agent = LangGraphAgent(
+math_agent = create_react_agent(
     name="math",
-    description="Specialist agent for math questions",
-    model=llm,
-    system_prompt="You provide help with math problems. Explain your reasoning at each step and include examples. \
+    tools=[],
+    model=model,
+    prompt="You provide help with math problems. Explain your reasoning at each step and include examples. \
         If prompted for anything else you refuse to answer.",
-    # tool_functions=[math_tool],
 )
 
-history_agent = LangGraphAgent(
+history_agent = CustomLangGraphAgent(
     name="history",
     description="Specialist agent for historical questions",
-    model=llm,
+    model=model,    
     system_prompt="You provide assistance with historical queries. Explain important events and context clearly.",
-)
+).graph
 
-triage_agent = LangGraphAgent(
-    name="GeneralAgent",
-    description="You determine which agent to use based on the user's question.",
-    model=llm,
-    system_prompt="You determine which agent to use based on the user's question.",
-    handoffs=[history_agent, math_agent],
-    checkpointer=session.checkpointer
-)
+supervisor = create_supervisor(
+    model=model,
+    agents=[history_agent, math_agent],
+    prompt=(
+        "You are a supervisor managing two agents:\n"
+        "- a history agent. Assign history-related tasks to this agent\n"
+        "- a math agent. Assign math-related tasks to this agent\n"
+        "Assign work to one agent at a time, do not call agents in parallel.\n"
+        "Do not do any work yourself."
+    )
+).compile(name="supervisor", checkpointer=session.checkpointer)
 
-AgentModule([triage_agent.graph, math_agent.graph, history_agent.graph])
+AgentModule([supervisor, history_agent, math_agent])
 
 if __name__ == "__main__":
     CLI.main()
