+++ examples/cli-langgraph-agents/pyproject.toml
@@ -8,6 +8,10 @@ requires-python = ">=3.12"
 dependencies = [
     "ak>=0.1.0",
     "ak-langgraph>=0.1.0",
+    "langchain-litellm>=0.2.2",
+    "langchain-openai>=0.3.28",
+    "langgraph-supervisor>=0.0.29",
+    "openai>=1.99.1",
 ]
 
 [tool.uv]
